Moving source files into `src` folder
