+++ CHANGELOG.md
@@ -1,0 +1,10 @@
+# v1.0.0
+  - Initial public introduction of PluMA to the public.
+  - Subsequent untagged commits are retroactively part of the v1.0.0 line.
+
+# v2.0.0
+  - Reworking of project build system
+  - Moving from using `cpp` suffix for C++ files for `cxx`.
+  - Consolidation of source code into `src` subdirectory.
+  - Adding Dockerfile and enabling generation of PluMA in a docker image.
+  -
