+++ src/PluGen/PluginGenerator.cxx
