+++ src/PluginManager.h
