+++ src/main.cxx
@@ -61,8 +61,6 @@ std::string toString(int val) {
 }
 //////////////////////////////////////////
 
-
-
 int main(int argc, char** argv)
 {
    /////////////////////////////////////////////////////////////////////////////////////////////////////////////
